jwt refresh works + delete

+++ src/main/java/com/example/demo/repositories/RefreshTokenRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByToken(String token);
 
-    int  deleteByBenutzer(Benutzer benutzer);
+    Long deleteByBenutzer(Benutzer benutzer);
 }
